+++ src/auth/auth.module.ts
@@ -2,7 +2,7 @@ import { Module } from '@nestjs/common';
 import { AuthService } from './auth.service';
 import { AuthController } from './auth.controller';
 import { UserModule } from '../user/user.module';
-import { ValidateCredentialConstraint } from './decorator/validator/validate-credential.validator';
+import { ValidateCredentialConstraint } from './validator/validate-credential.validator';
 
 @Module({
   imports: [UserModule],
